requirements.txt
colorama
cycler
fonttools
future
iso8601
joblib
kiwisolver
matplotlib
numpy
packaging
Pillow
pyparsing
PyQt5
#PyQt5-Qt5
PyQt5-sip
pyqtgraph
pyserial
python-dateutil
PyYAML
scikit-learn
scipy
six
#sklearn
tabulate
threadpoolctl
tqdm
